Achievement, trophy and 'rank' annoucement formatting inline with how it used to be.
These now have the proper color and font formatting from the way they
used to be years ago (thank you IRC client logs for preserving
color/font coding), and the output for achievements and trophies are the
way they were also (individually list four or fewer, just list number of
if 5 or greater).

'Rankings' have been changed a bit, since we no longer have a point
system or have actual rankings, even though the clans page lists the
clans by number of ascensions (highest on top). And it's generally held
that the clan with the most ascensions is the winning clan. The output
is now like this:

  [R] Clan Teenage NetHack Ninja Turtles moves up to position #2 with 15
      ascensions.

  [R] Clan Teenage NetHack Ninja Turtles drops to position #3.

Not a rank, but their position along with the number of clan ascensions.