Add delay between announcements, enable announcement suppression.
This commit adds flood protection, a delay of one second between
multiple achievement/trophy/clan position announcements. Will prevent
excess flood and being kicked from the IRC server.

Doing a full database wipe and data reimport could sometimes cause the
bot to treat all of the 'new' achievements and trophies as actually new
and then announce them all at once (hence the delay from earlier). This
actually isn't a bad thing when testing, but not at all ideal if having
to do a database rebuild in the middle of the tournament. So add a new
setting to allow suppression of these announcements to be turned on/off
via tnntbotconf.py (defaults to 'True').