Fix: don't display dumplog links for startscummed games.
The game code by design does not produce a dumplog file. The scoreboard
knows this, and will not produce a dumplog link for those games. The bot
was not aware of this, and would produce a weblink for a non-existent
dumplog file. The bot now knows to not do this, and will instead state
that the game was startscummed and that no dumplog exists.

Also in this commit - fixed an OLD bug from when we forked beholder bot
to create this bot. The definition of a scummed game did not match the
scoreboard's definition. Meaning some legit games were being suppressed
by the bot, while certain startscummed games were announced. Oops.